update
render adapter, fix bug, send intent to next activity

+++ app/src/main/java/com/example/eventsconferencespj/Activities/Home/ConfeData.kt
@@ -7,6 +7,6 @@ data class ConfeData(
     val seat: Int,
     val price: Int,
     val requiredFirstPay: Int,
-    val ratingStar: Int,
+    val ratingStar: Double,
     val image: Int
 )
